sync

+++ app/src/main/java/com/slash/youth/http/protocol/DeleteMyPublishTaskItemProtocol.java
@@ -12,10 +12,10 @@ import org.xutils.http.RequestParams;
  */
 public class DeleteMyPublishTaskItemProtocol extends BaseProtocol<SetBean> {
     private int type;
-    private int tid;
+    private long tid;
 
 
-    public DeleteMyPublishTaskItemProtocol(int type, int tid) {
+    public DeleteMyPublishTaskItemProtocol(int type, long tid) {
         this.type = type;
         this.tid = tid;
     }
