+++ app/src/main/java/com/slash/youth/ui/activity/MySkillManageActivity.java
@@ -80,7 +80,7 @@ public class MySkillManageActivity extends Activity implements View.OnClickListe
         }
     }
 
-    public void jumpMyAddSkillActivity(Activity activity,int skillListId) {
+    public void jumpMyAddSkillActivity(Activity activity,long skillListId) {
         Intent intentMyAddSkillActivity = new Intent(CommonUtils.getContext(), MyAddSkillActivity.class);
         intentMyAddSkillActivity.putExtra("skillListId",skillListId);
         activity.startActivityForResult(intentMyAddSkillActivity, Constants.SKILL_MANAGER_ONE);
